+++ src/commonMain/kotlin/kotools/types/experimental/Bound.kt
@@ -51,6 +51,13 @@ private constructor(override val value: T) : Bound<T> {
 @ExperimentalKotoolsTypesApi
 @ExperimentalSince(KotoolsTypesVersion.Unreleased)
 public class ExclusiveBound<out T : Comparable<@UnsafeVariance T>>
-internal constructor(override val value: T) : Bound<T> {
+private constructor(override val value: T) : Bound<T> {
     override fun toString(): String = "$value"
+
+    /** Contains static declarations for the [ExclusiveBound] type. */
+    public companion object {
+        @JvmSynthetic
+        internal infix fun <T : Comparable<T>> of(value: T): ExclusiveBound<T> =
+            ExclusiveBound(value)
+    }
 }